requirements.txt
openai>=1.40.0
chromadb>=0.5.0
pypdf>=4.2.0
python-dotenv>=1.0.1
numpy>=1.26.0
python-docx>=1.1.0
python-dotenv>=1.0.0